+++ package.json
@@ -4,6 +4,7 @@
   "description": "This is a dummy application used in testing the functionality of Konsult.",
   "main": "index.js",
   "scripts": {
+    "start": "node app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
